Stop hammering ir_translation id sequence
We've had some instances actually reaching the 2**31 limit, because
the temporary table used to prepare potential insertions used the same
id sequence as the main table, hence each insert in the temporary table
bumps the sequence.

Aggravations:

- if one hits the wall, the instance won't start
- if one resets the sequence too close from an existing id, it won't start either
- the existing ids are spread across the whole spectrum, making it hard to find
  a safe range to reset the sequence.

Yes this probably means instances that restart too much, but that's another issue.

The whole system is questionable at best [1] but this fix should have no impact, and
improve a bit startup performance and PostgreSQL health. On the other hand, we can't
guarantee without a long analysis that's out of our scope that even changing
ir_translation to a bigint would have no impact, yet alone give this the full rewrite
it certainly deserves.

[1] for starters, why use the DB at all in preparation, and use a temporary table
    that's logically tied to a specific connection ?